LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/parsing/border-image-source-computed.sub-expected.txt


PASS Property border-image-source value 'none' 
PASS Property border-image-source value 'url("http://localhost/")' 
FAIL Property border-image-source value 'linear-gradient(-45deg, red, currentcolor)' assert_equals: expected "linear-gradient(-45deg, rgb(255, 0, 0), rgb(0, 0, 255))" but got "linear-gradient(-45deg, red, currentcolor)"
FAIL Property border-image-source value 'repeating-linear-gradient(-45deg, red, 30%, currentcolor 70%, lime)' assert_equals: expected "repeating-linear-gradient(-45deg, rgb(255, 0, 0), 30%, rgb(0, 0, 255) 70%, rgb(0, 255, 0))" but got "repeating-linear-gradient(-45deg, red, 30%, currentcolor 70%, lime)"
FAIL Property border-image-source value 'radial-gradient(10px at 20px 30px, currentcolor, lime)' assert_equals: expected "radial-gradient(10px at 20px 30px, rgb(0, 0, 255), rgb(0, 255, 0))" but got "radial-gradient(10px at 20px 30px, currentcolor, lime)"
FAIL Property border-image-source value 'conic-gradient(from 90deg at 80% 90%, lime, black)' assert_equals: expected "conic-gradient(from 90deg at 80% 90%, rgb(0, 255, 0), rgb(0, 0, 0))" but got "conic-gradient(from 90deg at 80% 90%, lime, black)"
PASS url values are made absolute 
